requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==48.0.0
distro==1.9.0
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
google-auth==2.52.0
google-genai==2.0.1
gunicorn==26.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
MarkupSafe==3.0.3
openai==2.36.0
packaging==26.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
PyJWT==2.12.1
pymongo==4.17.0
pypdf==6.11.0
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
websockets==16.0
Werkzeug==3.1.8
